requirements.txt
# requirements.txt - List of required dependencies

# fastapi
# uvicorn
# torch>=2.0.0
# transformers>=4.35.0
# pillow
# huggingface_hub
# python-dotenv
# numpy>=1.24.0
# peft>=0.9.0
# bitsandbytes
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
bitsandbytes==0.45.5
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
peft==0.15.1
pillow==11.1.0
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0